refactor: node search component
