+++ src/app/view/editor-search-view/editor-node-search-view-component/editor-node-search-view-component.html
@@ -1,0 +1,33 @@
+<sbb-search
+  (search)="search()"
+  aria-label="{{ 'app.view.editor-search-view-component.search-node' | translate }}"
+>
+  <input
+    sbbInput
+    [placeholder]="'app.view.editor-search-view-component.search-node' | translate"
+    [formControl]="searchControl"
+    [sbbAutocomplete]="auto"
+  />
+</sbb-search>
+<sbb-autocomplete #auto="sbbAutocomplete" [displayWith]="displayNode">
+  @for (option of filteredNodes; track option) {
+  <sbb-option [value]="option">{{ transformNodeName(option) }}</sbb-option>
+  }
+</sbb-autocomplete>
+
+@if (searchResults.length > 1) {
+<div class="search-results-list">
+  @for (node of searchResults; track node) {
+  <br />
+  <p class="alignleft" style="line-height: 10px">
+    <span
+      class="smallstation"
+      [title]="transformNodeName(node)"
+      (click)="onSearchResultClick(node)"
+    >
+      {{ getNodeSearchValue(node) }}
+    </span>
+  </p>
+  }
+</div>
+}
