+++ src/app/view/editor-search-view/editor-search-view-component.scss
@@ -1,29 +1,3 @@
 @use "../../../variables" as *;
 @use "./../../view/editor-filter-view/editor-filter-view.component.scss";
 @use "../rastering/definitions" as *;
-
-.search-results-list {
-  margin-top: 0.5rem;
-  max-height: 200px;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.alignleft {
-  float: left;
-  width: 100%;
-  margin: 0;
-}
-
-span.smallstation {
-  cursor: pointer;
-  font-size: 12px;
-  border: 3px solid var(--sbb-color-background);
-  font-weight: lighter;
-  user-select: none;
-}
-
-span.smallstation:hover {
-  font-weight: bolder;
-  color: var(--COLOR_Edit);
-}
